+++ src/linkdialog.h
@@ -17,6 +17,8 @@ public:
 
     QString getSelectedNoteName();
     QString getURL();
+    QString getTitleForUrl(QUrl url);
+
 private slots:
     void on_searchLineEdit_textChanged(const QString &arg1);
     void on_notesListWidget_doubleClicked(const QModelIndex &index);