+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.56
 - added a shortcut list to the help menu and to the startup notes
+- you can now also link to URLs with `Ctrl + L` (or `Cmd + L` on OSX), if possible the title of the webpage will be fetched and used as link text
 
 ## 0.55
 - added some different shortcuts for the note history on the mac